Consolidated the two commands into one command and made necessary adjustments to all files.
