+++ src/main/java/frc/robot/RobotContainer.java
@@ -14,8 +14,7 @@ import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.Commands;
 import edu.wpi.first.wpilibj2.command.button.CommandXboxController;
 import edu.wpi.first.wpilibj2.command.sysid.SysIdRoutine.Direction;
-import frc.robot.commands.DecreaseElevatorLevelCommand;
-import frc.robot.commands.IncreaseElevatorLevelCommand;
+import frc.robot.commands.ElevatorLevelCommand;
 import frc.robot.generated.TunerConstants;
 import frc.robot.subsystems.CommandSwerveDrivetrain;
 import frc.robot.subsystems.ElevatorSubsystem;
@@ -73,8 +72,8 @@ public class RobotContainer {
         driver.leftBumper().onTrue(drivetrain.runOnce(() -> drivetrain.seedFieldCentric()));
 
         // Sets the pov up and down of operator's controller to increase or decrease the level of the arm elevator.
-        operator.povUp().onTrue(new IncreaseElevatorLevelCommand(elevatorSubsystem));
-        operator.povDown().onTrue(new DecreaseElevatorLevelCommand(elevatorSubsystem));
+        operator.povUp().onTrue(new ElevatorLevelCommand(elevatorSubsystem, true));
+        operator.povDown().onTrue(new ElevatorLevelCommand(elevatorSubsystem, false));
 
         drivetrain.registerTelemetry(logger::telemeterize);
     }
